+++ app/src/main/java/com/cbo/cbomobilereporting/ui_new/dcr_activities/StockistCall.java
@@ -76,6 +76,9 @@ import utils.adapterutils.SpinnerModel;
 import com.cbo.cbomobilereporting.MyCustumApplication;
 import com.uenics.javed.CBOLibrary.ResponseBuilder;
 
+import org.json.JSONArray;
+import org.json.JSONObject;
+
 import utils.networkUtil.NetworkUtil;
 import utils_new.AppAlert;
 import utils_new.Chemist_Gift_Dialog;
@@ -845,12 +848,21 @@ public class StockistCall extends AppCompatActivity implements ExpandableListAda
                                 .setResponse(new CBOServices.APIResponse() {
                                     @Override
                                     public void onComplete(Bundle bundle) throws Exception {
-                                        mstockistCall = (mStockistCall) new mStockistCall().setId(Dr_id);
-                                        stockistCallDB.delete(mstockistCall);
+                                        String table0 = bundle.getString("Tables0");
+                                        JSONArray jsonArray1 = new JSONArray(table0);
+                                        JSONObject object = jsonArray1.getJSONObject(0);
 
-                                        cbohelp.delete_Stokist_from_local_all(Dr_id);
-                                        customVariablesAndMethod.msgBox(context,Dr_name+" sucessfully Deleted.");
-                                        finish();
+                                        if (object.getString("DCRID").equalsIgnoreCase("1")) {
+
+                                            mstockistCall = (mStockistCall) new mStockistCall().setId(Dr_id);
+                                            stockistCallDB.delete(mstockistCall);
+
+                                            cbohelp.delete_Stokist_from_local_all(Dr_id);
+                                            customVariablesAndMethod.msgBox(context,Dr_name+" sucessfully Deleted.");
+                                            finish();
+                                        }else{
+                                            AppAlert.getInstance().getAlert(context,"Alert",object.getString("DCRID"));
+                                        }
                                     }
 
                                     @Override
